indicators/requirements.txt
pandas
numpy
matplotlib
fastapi
uvicorn
psycopg2-binary
python-dotenv
requests
